+++ client/src/Pages/Posting/Post.js
@@ -87,7 +87,10 @@ const Post = ({ location, history }) => {
       }
       setDel(false);
     };
-    deletePost();
+
+    if (window.confirm("정말 삭제하시겠습니까?")) {
+      deletePost();
+    }
   };
 
   return (
